+++ NUTEE/src/main/java/kr/nutee/repository/mapper/ArticleEmpathyMapper.java
@@ -1,5 +1,6 @@
 package kr.nutee.repository.mapper;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
@@ -13,15 +14,15 @@ import kr.nutee.dao.ArticleEmpathy;
 @Mapper
 public interface ArticleEmpathyMapper {
 	//article id로 검색
-	List<ArticleEmpathy> findAllByArticleId(int articleId);
+	List<ArticleEmpathy> findAllByArticleId(BigInteger articleId);
 
 	//게시글 별 공감 갯수 출력
-	int countByArticleId(int articleId);
+	BigInteger countByArticleId(BigInteger articleId);
 
 	//공감
-	void insert(int userId, int articleId);
+	void insert(long userId, BigInteger articleId);
 
 	//공감 취소
-	void delete(int userId, int articleId);
+	void delete(long userId, BigInteger articleId);
 
 }
