+++ NUTEE/src/main/java/kr/nutee/service/ArticleEmpathyService.java
@@ -1,5 +1,6 @@
 package kr.nutee.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import kr.nutee.dao.ArticleEmpathy;
@@ -12,15 +13,15 @@ import kr.nutee.dao.ArticleEmpathy;
 public interface ArticleEmpathyService {
 
 	//article id로 검색
-	List<ArticleEmpathy> findAllByArticleId(int articleId);
+	List<ArticleEmpathy> findAllByArticleId(BigInteger articleId);
 
 	//게시글 별 공감 갯수 출력
-	int countByArticleId(int articleId);
+	BigInteger countByArticleId(BigInteger articleId);
 
 	//공감
-	void insert(int userId, int articleId);
+	void insert(long userId, BigInteger articleId);
 
 	//공감 취소
-	void delete(int userId, int articleId);
+	void delete(long userId, BigInteger articleId);
 
 }
